rm build.py
